+++ libft/double_linked_list/ft_dlist_sort.c
@@ -6,13 +6,13 @@
 /*   By: nharra <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/10 11:12:55 by nharra            #+#    #+#             */
-/*   Updated: 2019/10/11 18:27:34 by nharra           ###   ########.fr       */
+/*   Updated: 2019/10/11 19:32:47 by nharra           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_dlist_sort(t_dlist *beg, long (*cmp)(const void *, const void *))
+void	ft_dlist_sort(t_dlist *beg, int (*cmp)(const void *, const void *))
 {
 	t_dlist *ptr;
 	t_dlist *min;
